refactor: move pattern grid code to separate package
