+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/autocrafting/patterngrid/Alternative.java
@@ -1,4 +1,4 @@
-package com.refinedmods.refinedstorage.common.autocrafting;
+package com.refinedmods.refinedstorage.common.autocrafting.patterngrid;
 
 import com.refinedmods.refinedstorage.common.api.support.resource.PlatformResourceKey;
